STYLE: Synchronize with latest gdcm 0.6 branch, mostly comments

+++ Utilities/gdcm/src/gdcmDocument.cxx
@@ -1320,7 +1320,7 @@ void Document::ParseDES(DocEntrySet *set, long offset,
    
    while (true)
    { 
-      if ( !delim_mode && ((long)Fp->tellg()-offset) >= l_max)
+      if ( !delim_mode && ((long)Fp->tellg()-offset) >= l_max )
       {
          break;
       }
@@ -1363,7 +1363,7 @@ void Document::ParseDES(DocEntrySet *set, long offset,
             {
                break;
             }
-            if ( !delim_mode && ((long)Fp->tellg()-offset) >= l_max)
+            if ( !delim_mode && ((long)Fp->tellg()-offset) >= l_max )
             {
                break;
             }
@@ -1473,7 +1473,7 @@ void Document::ParseDES(DocEntrySet *set, long offset,
                      l, delim_mode);
          }
          set->AddEntry( newSeqEntry );
-         if ( !delim_mode && ((long)Fp->tellg()-offset) >= l_max)
+         if ( !delim_mode && ((long)Fp->tellg()-offset) >= l_max )
          {
             break;
          }
@@ -1508,7 +1508,7 @@ void Document::ParseSQ( SeqEntry* seqEntry,
             break;
          }
       }
-      if ( !delim_mode && ((long)Fp->tellg()-offset) >= l_max)
+      if ( !delim_mode && ((long)Fp->tellg()-offset) >= l_max )
       {
           break;
       }
@@ -1756,7 +1756,7 @@ void Document::FindDocEntryLength( DocEntry *entry )
                // the length and proceed.
                long currentPosition = Fp->tellg();
                Fp->seekg(0L,std::ios::end);
-               long lengthUntilEOF = (long)Fp->tellg() - currentPosition;
+               long lengthUntilEOF = Fp->tellg() - currentPosition;
                Fp->seekg(currentPosition, std::ios::beg);
                entry->SetLength(lengthUntilEOF);
                return;
